+++ spec/lib/virtual_assembly/semantizer/semantic_object_spec.rb
@@ -6,6 +6,11 @@ describe VirtualAssembly::Semantizer::SemanticObject do
       include VirtualAssembly::Semantizer::SemanticObject
     end.new
   end
+  let(:other_clazz) do
+    Class.new(Object) do
+      include VirtualAssembly::Semantizer::SemanticObject
+    end
+  end
 
   it 'has a semantic id' do
     expect { object.semanticId = 'five' }
@@ -26,5 +31,21 @@ describe VirtualAssembly::Semantizer::SemanticObject do
 
     expect { object.semanticProperty('@id').value = 'six' }
       .to change { object.semanticPropertyValue('@id') }.from('five').to('six')
+  end
+
+  it 'can be compared' do
+    expect(object).to eq object.clone
+    expect(object).to eq other_clazz.new
+
+    expect(object).to_not eq other_clazz.new("id5")
+    expect(object).to_not eq other_clazz.new(nil, "typeA")
+
+    object.semanticId = "id5"
+    object.semanticType = "typeA"
+
+    expect(object).to eq other_clazz.new("id5", "typeA")
+
+    expect(object).to_not eq other_clazz.new("id5")
+    expect(object).to_not eq other_clazz.new(nil, "typeA")
   end
 end